more functions added

+++ odooenv.py
@@ -2,6 +2,23 @@
 import argparse
 import subprocess
 
+ODOOVER = '8.0'
+HOME = '~/odoo-test-' + ODOOVER + '/'
+PSQL = '~/postgresql-test/'
+
+ODOO = 'jobiols/odoo-adhoc:' + ODOOVER
+AEROO = "jobiols/aeroo-docs"
+POSTGRES = "postgres:9.4"
+BACKUP = "jobiols/backup"
+
+IMAGES = [ODOO, AEROO, POSTGRES, BACKUP]
+CLIENTS = [{'client': 'str', 'port': '8069'},
+           {'client': 'makeover', 'port': '8070'}]
+
+REPOS = ['odoo-addons', 'odoo-argentina', 'aeroo_reports', 'server-tools', 'web', 'management-system', 'knowledge',
+         'str']
+
+
 # ###############################
 RED = "\033[1;31m"
 GREEN = "\033[1;32m"
@@ -25,15 +42,9 @@ def red(string):
 def yellow_light(string):
     return YELLOW_LIGHT + string + CLEAR
 
+
 ###################################
 
-ODOOVER = '8.0'
-HOME = '~/odoo-test-' + ODOOVER + '/'
-PSQL = '~/postgresql-test/'
-REPOS = ['odoo-addons', 'odoo-argentina', 'aeroo_reports', 'server-tools', 'web', 'management-system', 'knowledge',
-         'str']
-ODOO = 'jobiols/odoo-adhoc:' + ODOOVER
-CLIENTS = [{'client': 'str', 'port': '8069'}]
 
 
 def msgrun(msg):
@@ -48,10 +59,23 @@ def msgerr(msg):
     print red(msg)
 
 
+def msginf(msg):
+    print yellow(msg)
+
+
+def client_port(client):
+    port = ''
+    for cli in CLIENTS:
+        if cli['client'] == client:
+            port = cli['port']
+    return port
+
+
 def uninstall_environment():
     msgrun('Uninstalling odoo environment')
     subprocess.call(['sudo rm -r ' + HOME], shell=True)
-    subprocess.call(['sudo rm -r ' + PSQL], shell=True)
+    if raw_input('Delete postgresql directory? (y/n) ') == 'y':
+        subprocess.call(['sudo rm -r ' + PSQL], shell=True)
     return True
 
 
@@ -65,17 +89,23 @@ def install_environment():
     # install sources
     subprocess.call(['mkdir -p ' + HOME + 'sources'], shell=True)
 
+
     # subprocess.call(['chmod 777 -R '+HOME+'/sources'], shell=True)
+
     for repo in REPOS:
-        msgrun('pulling repo ' + repo)
-        subprocess.call(
-            'git clone -b ' + ODOOVER + ' --depth 1 http://github.com/jobiols/' + repo + ' ' + HOME + 'sources/' + repo,
-            shell=True)
+        msginf('pulling repo ' + repo)
+        if subprocess.call('git clone -b ' + ODOOVER +
+                                   ' --depth 1 http://github.com/jobiols/' + repo + ' ' + HOME + 'sources/' + repo,
+                           shell=True):
+            msgerr('Fail installing environment, uninstall and try again.')
+            return True
+
     msgdone('Install Done')
     return True
 
 
 def install_client():
+    msgrun('Install clients')
     # Calculate addon_path
     path = '/mnt/extra-addons/'
     addon_path = path + REPOS[0]
@@ -85,64 +115,116 @@ def install_client():
     for cli in args.client:
         msgrun('Installing Odoo image for client ' + cli)
         # Creating directory's for client
-        subprocess.call('mkdir -p ' + HOME + cli + '/config')
-        subprocess.call('mkdir -p ' + HOME + cli + '/data_dir')
+        subprocess.call('mkdir -p ' + HOME + cli + '/config', shell='True')
+        subprocess.call('mkdir -p ' + HOME + cli + '/data_dir', shell='True')
+        subprocess.call(['chmod 777 -R ' + HOME + cli], shell=True)
 
         # creating config file for client
         subprocess.call(
-            'docker run --rm \
+            'sudo docker run --rm \
             -v ' + HOME + cli + '/config:/etc/odoo \
             -v ' + HOME + 'sources:/mnt/extra-addons \
             -v ' + HOME + cli + '/data_dir:/var/lib/odoo \
-            --name ' + cli +
+            --name ' + cli + ' ' + \
             ODOO + ' -- --stop-after-init -s \
             --addons-path=' + addon_path, shell=True)
-
     msgdone('Installing done')
     return True
 
 
-def run_environment():
-    msgrun("Running AEROO image")
+def run_aeroo_image():
+    msgrun("Running aeroo and postgres images")
     subprocess.call(
-        'docker run -d \
+        'sudo docker run -d \
 	    -p 127.0.0.1:8989:8989 \
 	    --name="aeroo_docs" \
-    	--restart=always \
-	    ${AEROO} ')
-    msgdone("AEROO up and running")
+    	--restart=always ' + \
+        AEROO, shell=True)
+    return True
 
-    msgrun ("Running POSTGRES")
+
+def run_environment():
+    run_aeroo_image()
+
     subprocess.call(
-        'docker run -d \
+        'sudo docker run -d \
 	    -e POSTGRES_USER=odoo \
 	    -e POSTGRES_PASSWORD=odoo \
 	    -v ~/postgresql:/var/lib/postgresql/data \
 	    --restart=always \
-	    --name db-odoo \
-	    ${POSTGRES} ')
- 	msgdone ("POSTGRES up and running")
+	    --name db-odoo ' + \
+        POSTGRES, shell=True)
+    msgdone("Images up and running")
 
     return True
 
 
 def run_developer():
+    msgrun('Running developer mode')
+    run_aeroo_image()
+
+    subprocess.call(
+        'sudo docker run -d \
+	    -e POSTGRES_USER=odoo \
+        -p 5432:5432 \
+	    -e POSTGRES_PASSWORD=odoo \
+	    -v ~/postgresql:/var/lib/postgresql/data \
+	    --restart=always \
+	    --name db-odoo ' + \
+        POSTGRES, shell=True)
+
     return True
 
 
 def run_client():
-    return True
+    for cli in args.client:
+        msgrun('starting ' + cli)
+        subprocess.call(
+            'sudo docker run -d \
+            --link aeroo_docs:aeroo \
+            -p ' + client_port(cli) + ':8069 \
+            -v ~/odoo80/' + cli + '/config:/etc/odoo \
+            -v ~/odoo80/' + cli + '/data_dir:/var/lib/odoo \
+            -v ~/odoo80/sources:/mnt/extra-addons \
+            --link db-odoo:db \
+            --restart=always \
+            --name ' + cli + ' ' + \
+            ODOO + ' -- --db-filter= ' + cli, shell=True)
 
-
-def stop_environment():
     return True
 
 
 def stop_client():
+    msgrun('stopping clients ' + ', '.join(args.client))
+    for cli in args.client:
+        msgrun('stop image for client ' + cli)
+        subprocess.call('sudo docker stop ' + cli, shell=True)
+        subprocess.call('sudo docker rm ' + cli, shell=True)
+
+    msgdone('all client stopped')
+
+    return True
+
+
+def stop_environment():
+    msgrun('Stopping environment')
+
+    for name in ['db-odoo', 'aeroo_docs']:
+        msgrun
+        'Stoping image ' + name
+        subprocess.call('sudo docker stop ' + name, shell=True)
+        subprocess.call('sudo docker rm ' + name, shell=True)
+
     return True
 
 
 def pull_all_images():
+    msgrun('Pulling all images for ' + ODOOVER)
+
+    for image in IMAGES:
+        msgrun('Pulling ' + image)
+        subprocess.call('sudo docker pull ' + image, shell=True)
+
     return True
 
 
@@ -158,10 +240,38 @@ def list_data():
 
 
 def no_ip_install():
+    msgrun('Installing no-ip client')
+    subprocess.call('apt-get install make', shell=True)
+    subprocess.call('apt-get -y install gcc', shell=True)
+
+    """
+        cd /usr/local/src/
+        wget http://www.noip.com/client/linux/noip-duc-linux.tar.gz
+        tar xf noip-duc-linux.tar.gz
+        cd noip-2.1.9-1/
+        make install
+        msginfo "Please answer some questions"
+        #Please enter login/email: jobiols
+        #Please enter password: veconceR
+        #contestar preguntas sobre defaults y grupos.
+        rm /usr/local/src/noip-duc-linux.tar.gz
+        #Poner el script en init.d
+        cp /usr/local/src/noip-2.1.9-1/debian.noip2.sh  /etc/init.d/
+        #Make script executable
+        chmod +x /etc/init.d/debian.noip2.sh
+        #hacer que se ejecute al booteo
+        update-rc.d debian.noip2.sh defaults
+        #arrancar y parar el demonio
+        msgrun "Starting no-ip daemon"
+        /etc/init.d/debian.noip2.sh restart
+        msgdone "no-ip service running"
+    """
     return True
 
 
 def docker_install():
+    msgrun('Installing docker')
+    subprocess.call('wget -qO- https://get.docker.com/ | sh', shell=True)
     return True
 
 
@@ -192,16 +302,27 @@ if __name__ == '__main__':
 
     args = parser.parse_args()
 
-    if args.uninstall_env:        uninstall_environment()
-    if args.install_env:         install_environment()
-    if args.install_cli:        install_client()
-    if args.run_env:        run_environment()
-    if args.run_dev: run_developer()
-    if args.run_cli: run_client()
-    if args.stop_env: stop_environment()
-    if args.stop_cli: stop_client()
-    if args.pull_all: pull_all_images()
-    if args.list: list_data()
-    if args.no_ip_install: no_ip_install()
-    if args.docker_install: docker_install()
-
+    if args.uninstall_env:
+        uninstall_environment()
+    if args.install_env:
+        install_environment()
+    if args.install_cli:
+        install_client()
+    if args.run_env:
+        run_environment()
+    if args.run_dev:
+        run_developer()
+    if args.run_cli:
+        run_client()
+    if args.stop_env:
+        stop_environment()
+    if args.stop_cli:
+        stop_client()
+    if args.pull_all:
+        pull_all_images()
+    if args.list:
+        list_data()
+    if args.no_ip_install:
+        no_ip_install()
+    if args.docker_install:
+        docker_install()
